requirements.txt
greenlet==2.0.2
maskpass==0.3.7
numpy==1.24.2
pandas==1.5.3
psycopg2==2.9.5
pycodestyle==2.10.0
pynput==1.7.6
pyobjc-core==9.0.1
pyobjc-framework-ApplicationServices==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Quartz==9.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
